% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_P.R
\name{cond}
\alias{cond}
\title{\code{cond} helper function}
\usage{
cond(l, r_breaks, sd)
}
\arguments{
\item{l}{the given value of the latent variable}

\item{r_breaks}{a (n+1) vector defining a partition of the interval R into n sets}

\item{sd}{the standard deviation of the half-normal distribution}
}
\value{
an n vector whose i^th entry is P_{ij} as defined above.
}
\description{
'cond' computes the inner integral in () wrt the observed coordinate \code{r} for a given the latent coordinate \code{l}.
This is a helper function for \code{example_P} and should not be used independently.
}
