% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_exp.R
\name{cond_exp}
\alias{cond_exp}
\title{\code{cond_exp} helper function}
\usage{
cond_exp(l, r_breaks, rate)
}
\arguments{
\item{l}{the given value of the latent variable}

\item{r_breaks}{a (n+1) vector defining a partition of the interval R into n sets}

\item{rate}{the rate of the exponential distribution}
}
\value{
an n vector whose i^th entry is P_{ij} as defined above.
}
\description{
'cond_exp' computes the inner integral in () wrt the observed coordinate \code{r} for a given the latent coordinate \code{l}.
This is a helper function for \code{P_exp} and should not be used independently.
}
