% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_hnorm.R
\name{example_lambda}
\alias{example_lambda}
\title{Intensity function for example NHPP}
\usage{
example_lambda(x, height)
}
\arguments{
\item{x}{numeric n-vector giving coordinates to evaluate the intensity}

\item{height}{scaling parameter equal to expected number of observations over [0,100]}
}
\value{
n-vector giving the value of the intensity
}
\description{
Intensity function for example NHPP
}
\examples{
# Evaluate example_lambda over a uniform grid
l_grid <- seq(0,100, length.out = 20)
height <- 25000
example_lambda(l_grid, height)
}
