% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_P.R
\name{example_P}
\alias{example_P}
\title{P matrix for half-normal example}
\usage{
example_P(l_breaks, r_breaks, sd, n_quad = 100)
}
\arguments{
\item{l_breaks}{a (m+1) vector defining a partition of the interval L into m sets}

\item{r_breaks}{a (n+1) vector defining a partition of the interval R into n sets}

\item{sd}{the standard deviation of the half-normal distribution}

\item{n_quad}{the number of quadrature points for integration over interval L}
}
\value{
the n x m matrix P of transition probabilities
}
\description{
P matrix for half-normal example
}
\examples{
# Compute P for half-normal measurement error with sd = 2
# The latent interval L = [0,10] is divided into 20 intervals of length 0.5
# The observed interval R = [0,15] is divided into 15 intervals of length 1.
l_breaks <- seq(0,10, length.out = 21)
r_breaks <- seq(0,15, length.out = 16)
P <- example_P(l_breaks, r_breaks, sd = 2, n_quad = 100)
}
