% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{gradient.ppdeconvFix}
\alias{gradient.ppdeconvFix}
\title{gradient for fixed P ppdeconv models}
\usage{
\method{gradient}{ppdeconvFix}(obj)
}
\arguments{
\item{obj}{a \code{ppdeconvFix} object}
}
\value{
the gradient of the penalized log-likelihood
}
\description{
gradient for fixed P ppdeconv models
}
\examples{
#TODO
}
