% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{gradient.ppdeconvObj}
\alias{gradient.ppdeconvObj}
\title{gradient for fixed P ppdeconv models}
\usage{
\method{gradient}{ppdeconvObj}(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvObj} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{
the gradient of the penalized log-likelihood
}
\description{
gradient for fixed P ppdeconv models
}
\examples{
#TODO
}
