% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{gradient}
\alias{gradient}
\title{gradient for ppdeconv models}
\usage{
gradient(x)
}
\arguments{
\item{x}{a \code{ppdeconvFix} or \code{ppdeconvVar} object}
}
\value{
the gradient of the penalized log-likelihood
}
\description{
TODO: Add more detail
}
\examples{
#TODO
}
