% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{hessian.ppdeconvObj}
\alias{hessian.ppdeconvObj}
\title{hessian for the penalized log-likelihood}
\usage{
\method{hessian}{ppdeconvObj}(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvObj} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{
the hessian of the penalized log-likelihood
}
\description{
hessian for the penalized log-likelihood
}
\examples{
#TODO
}
