% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{hessian}
\alias{hessian}
\title{title}
\usage{
hessian(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{

}
\description{
title
}
\examples{
#TODO
}
