% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{lddot.ppdeconvFix}
\alias{lddot.ppdeconvFix}
\title{hessian of log-likelihood for \code{ppdeconvFix}}
\usage{
\method{lddot}{ppdeconvFix}(x)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}
}
\value{
the hessian of the log-likelihood
}
\description{
hessian of log-likelihood for \code{ppdeconvFix}
}
\examples{
#TODO
}
