% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{lddot.ppdeconvObj}
\alias{lddot.ppdeconvObj}
\title{hessian of log-likelihood for \code{ppdeconvObj}}
\usage{
\method{lddot}{ppdeconvObj}(x, a)
}
\arguments{
\item{x, }{a a \code{ppdeconvObj} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{
the hessian of the log-likelihood
}
\description{
hessian of log-likelihood for \code{ppdeconvObj}
}
\examples{
#TODO
}
