% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{ldot.default}
\alias{ldot.default}
\title{Title}
\usage{
\method{ldot}{default}(x, a)
}
\arguments{
\item{x}{any object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{

}
\description{
Title
}
\examples{
#TODO
}
