% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{ldot.ppdeconvFix}
\alias{ldot.ppdeconvFix}
\title{\code{ldot} for fixed P ppdeconv model}
\usage{
\method{ldot}{ppdeconvFix}(x)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}
}
\value{
the gradient of the log-likelihood
}
\description{
\code{ldot} for fixed P ppdeconv model
}
\examples{
#TODO
}
