% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{ldot}
\alias{ldot}
\title{gradient of log-likelihood}
\usage{
ldot(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvFix} or \code{ppdeconvVar} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{
the gradient of the log-likelihood
}
\description{
See documentation of \code{gradient} for details
}
\examples{
#TODO
}
