% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglik}
\alias{loglik}
\title{Title}
\usage{
loglik(a, y, Q, P, S, c0)
}
\arguments{
\item{a}{the p-vector initial guess for mle}

\item{y}{the n observed cell counts}

\item{Q}{the m x p structure matrix}

\item{P}{the n x m transition matrix#'}

\item{S}{the p x p penalty matrix}

\item{c0}{the scalar penalty}
}
\value{
the penalized log-likelihood
}
\description{
Computes the penalized log-likelihood for the mis-measured point process
with transition matrix P and latent structure matrix Q as described in
equations (24) and (25)
}
\examples{
# TODO
}
