% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglik_P}
\alias{loglik_P}
\title{log-likelihood for the ppdeconv model, variable P}
\usage{
loglik_P(a, y, Q, P_fn, S, c0)
}
\arguments{
\item{a}{the p-vector initial guess for mle}

\item{y}{the n observed cell counts}

\item{Q}{the m x p structure matrix}

\item{P_fn}{a one-parameter function that returns the n x m transition matrix}

\item{S}{the p x p penalty matrix}

\item{c0}{the scalar penalty}
}
\value{
the penalized log-likelihood
}
\description{
Computes the penalized log-likelihood for the mis-measured point process
with transition matrix P that varies with a parameter and latent structure
matrix Q as described in equations (23) and (25)
}
\examples{
# TODO
}
