% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_ppdeconvObj}
\alias{new_ppdeconvObj}
\title{ppdeconv Object}
\usage{
new_ppdeconvObj(
  N = numeric(),
  u = data.frame(),
  Q = matrix(nrow = 0, ncol = 0),
  P_fn,
  S = matrix(nrow = 0, ncol = 0),
  c0 = numeric(),
  l_breaks = numeric(),
  r_breaks = numeric(),
  l_grid = numeric(),
  r_grid = numeric(),
  a = numeric(),
  b = numeric()
)
}
\arguments{
\item{N}{a vector of length n of observed counts}

\item{u}{a data.frame of covariates with n ro}

\item{Q}{a m x p structure matrix}

\item{P_fn}{a function on \code{ppdeconvObj} objects that returns an
n x m matrix of transition probabilities}

\item{S}{a p x p penalty matrix}

\item{c0}{a numeric smoothness parameter}

\item{l_breaks}{an (m + 1)-vector of latent space break points}

\item{r_breaks}{an (n + 1)-vector of observed space breaks points}

\item{l_grid}{an m-vector of points between the breaks of \code{l_breaks}}

\item{r_grid}{an n-vector of points between the breaks of \code{r_breaks}}

\item{a}{a p-dimensional parameter vector corresponding to \code{Q}}

\item{b}{a q-dimensional parameter vector corresponding to \code{P}}
}
\value{
a \code{ppdeconvObj} object
}
\description{
ppdeconv Object
}
\examples{
#TODO
}
