% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_ppdeconvVar}
\alias{new_ppdeconvVar}
\title{Title}
\usage{
new_ppdeconvVar(
  N = numeric(),
  Q = matrix(nrow = 0, ncol = 0),
  P = matrix(nrow = 0, ncol = 0),
  S = matrix(nrow = 0, ncol = 0),
  c0 = numeric(),
  l_breaks = numeric(),
  r_breaks = numeric(),
  l_grid = numeric(),
  r_grid = numeric(),
  a = NULL,
  a_idx = NULL,
  b,
  b_idx = NULL,
  M = NULL,
  P_fn
)
}
\arguments{
\item{P_fn}{}
}
\value{

}
\description{
Title
}
