% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_exp.R
\name{P_exp}
\alias{P_exp}
\title{Title}
\usage{
P_exp(l_breaks, r_breaks, rate, n_quad = 1)
}
\arguments{
\item{l_breaks}{a (m+1) vector defining a partition of the interval L into m sets}

\item{r_breaks}{a (n+1) vector defining a partition of the interval R into n sets}

\item{rate}{the rate of the exponential distribution}
}
\value{
the n x m matrix P of transition probabilities
}
\description{
Title
}
\examples{
# Compute P for exponential measurement error with rate = 1
# The latent interval L = [0,10] is divided into 20 intervals of length 0.5
# The observed interval R = [0,15] is divided into 15 intervals of length 1.
l_breaks <- seq(0,10, length.out = 21)
r_breaks <- seq(0,15, length.out = 16)
P <- P_exp(l_breaks, r_breaks, rate = 1)
}
