% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_hnorm.R
\name{phnorm}
\alias{phnorm}
\title{CDF of Half-normal distribution}
\usage{
phnorm(q, mean = 0, sd)
}
\arguments{
\item{q}{vector of quantiles}

\item{mean}{scalar of mean}

\item{sd}{the standard deviation of the half-normal distribution}
}
\value{
the CDF for the given vector of quantiles \code{q}
}
\description{
CDF of Half-normal distribution
}
