% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppdeconv.R
\name{ppdeconv}
\alias{ppdeconv}
\title{Title}
\usage{
ppdeconv(x, a0 = NULL, method = "mle", control = list())
}
\arguments{
\item{x}{a list containing the data for \code{pdeconv}. See details.}

\item{a0}{an initial estimate for the parameter}

\item{mode}{a string indicating what type of inference to perform.}
}
\value{
a model fit object
}
\description{
Title
}
\examples{
# TODO
}
