% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{sdot.ppdeconvFix}
\alias{sdot.ppdeconvFix}
\title{\code{sdot} for fixed P ppdeconv model}
\usage{
\method{sdot}{ppdeconvFix}(x)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}
}
\value{

}
\description{
\code{sdot} for fixed P ppdeconv model
}
\examples{
#TODO
}
