% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{sdot.ppdeconvObj}
\alias{sdot.ppdeconvObj}
\title{\code{sdot} for fixed P ppdeconv model}
\usage{
\method{sdot}{ppdeconvObj}(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvObj} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{

}
\description{
\code{sdot} for fixed P ppdeconv model
}
\examples{
#TODO
}
