% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{sdot}
\alias{sdot}
\title{The gradient of the penalty term}
\usage{
sdot(a, y, Q, P, S, c0)
}
\arguments{
\item{a}{the p-dimensional parameter vector}

\item{y}{the n-dimensional data vector of counts}

\item{Q}{the m x p structure matrix}

\item{P}{the n x m matrix of transition probabilities}

\item{S}{the p x p penalty matrix}

\item{c0}{the smoothness parameter}
}
\value{
the p-dimensional gradient vector
}
\description{
The gradient of the penalty term
}
\examples{
#TODO
}
