% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_nhpp.R
\name{sim_nhpp}
\alias{sim_nhpp}
\title{Simulate non-homogeneous Poisson process on the real line}
\usage{
sim_nhpp(lambda, min, max, M = NULL)
}
\arguments{
\item{lambda}{non-negative intensity function. The first argument is the coordinate on the real line.}

\item{min}{lower limit of the observation interval}

\item{max}{upper limit of the observation interval}

\item{M}{an upper bound for the function lambda. If M = \code{NULL}, an upper bound is obtained using \code{optimize}.}
}
\value{
a vector containing the points of the realization
}
\description{
Simulate non-homogeneous Poisson process on the real line
}
\examples{
# Using known bound M over interval [0,10]
lambda <- function(x) abs(sin(x))
y <- sim_nhpp(lambda = lambda, min = 0, max = 10, M = 1)
# Using unknown bound M over interval [0,10]
y <- sim_nhpp(lambda = lambda, min = 0, max = 10, M = NULL)
}
