% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_bdiag}
\alias{to_bdiag}
\title{Put data into block diagonal format}
\usage{
to_bdiag(data)
}
\arguments{
\item{data}{list of data for the ppdeconv function}
}
\value{
list of data in block diagonal format
}
\description{
Put data into block diagonal format
}
\examples{
# TODO
}
