% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{gradient}
\alias{gradient}
\title{Compute gradient of the penalized log-likelihood}
\usage{
gradient(a, y, Q, P, S, c0)
}
\arguments{
\item{a}{the p-dimensional parameter vector}

\item{y}{the n-dimensional data vector of counts}

\item{Q}{the m x p structure matrix}

\item{P}{the n x m matrix of transition probabilities}

\item{S}{the p x p penalty matrix}

\item{c0}{the smoothness parameter}
}
\value{
the p-dimesional gradient vector
}
\description{
Compute gradient of the penalized log-likelihood
}
\examples{
#TODO
}
