% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{hessian.ppdeconvFix}
\alias{hessian.ppdeconvFix}
\title{hessian for the penalized log-likelihood}
\usage{
\method{hessian}{ppdeconvFix}(x)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}
}
\value{
the hessian of the penalized log-likelihood
}
\description{
hessian for the penalized log-likelihood
}
\examples{
#TODO
}
