% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{lddot.ppdeconvFix}
\alias{lddot.ppdeconvFix}
\title{hessian of log-likelihood for \code{ppdeconvFix}}
\usage{
\method{lddot}{ppdeconvFix}(x, a)
}
\arguments{
\item{x, }{a a \code{ppdeconvFix} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{
the hessian of the log-likelihood
}
\description{
hessian of log-likelihood for \code{ppdeconvFix}
}
\examples{
#TODO
}
