% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{lddot}
\alias{lddot}
\title{The Hessian of the unpenalized log-likelihood}
\usage{
lddot(a, y, Q, P, S, c0)
}
\arguments{
\item{a}{the p-dimensional parameter vector}

\item{y}{the n-dimensional data vector of counts}

\item{Q}{the m x p structure matrix}

\item{P}{the n x m matrix of transition probabilities}

\item{S}{the p x p penalty matrix}

\item{c0}{the smoothness parameter}
}
\value{
the p x p Hessian of the unpenalized log-likelihood with respect to the parameter \code{a}
}
\description{
The Hessian of the unpenalized log-likelihood
}
