% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglik}
\alias{loglik}
\title{log-likelihood for \code{ppdeconv} object}
\usage{
loglik(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}
}
\value{
the log-likelihood of the object
}
\description{
log-likelihood for \code{ppdeconv} object
}
\examples{
#TODO
}
