% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{sddot}
\alias{sddot}
\title{Title}
\usage{
sddot(x, a)

sddot(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvFix} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{


the hessian of the penalty term
}
\description{
Title

hessian for the penalty term of \code{ppdeconvFix} objects
}
\examples{
#TODO
#TODO
}
