% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivs.R
\name{sdot}
\alias{sdot}
\title{gradient of the penalty term}
\usage{
sdot(x, a)
}
\arguments{
\item{x}{a \code{ppdeconvFix} or \code{ppdeconvVar} object}

\item{a}{a parameter vector with length(a) = ncol(x$Q)}
}
\value{
the gradient of the penalty term
}
\description{
See documentation of \code{gradient} for details
}
\examples{
#TODO
}
