% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hpp.R
\name{sim_hpp}
\alias{sim_hpp}
\title{Simulate realization from a homogeneous Poisson process}
\usage{
sim_hpp(rate = 1, min = 0, max = 1)
}
\arguments{
\item{rate}{scalar giving the rate parameter of the process}

\item{min}{lower limit of the observation interval}

\item{max}{upper limit of the observation interval}
}
\value{
a vector containing the points of the realization within the given observation interval
}
\description{
Simulate realization from a homogeneous Poisson process
}
\examples{
# A realization from a Poisson process with rate 10 over the unit interval [0,1]
y <- sim_hpp(rate = 10, min = 0, max = 1)
}
